requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1 # uvicorn을 위한 표준 의존성 포함
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4 # 비밀번호 해싱을 위해 bcrypt 백엔드 포함
python-jose[cryptography]==3.3.0 # JWT 토큰 사용을 위해 cryptography 백엔드 포함
python-multipart==0.0.9 # 폼 데이터 처리를 위해
pydantic==2.7.4
pydantic-settings==2.3.3
qdrant-client==1.9.0
redis==5.0.4
celery==5.4.0
gunicorn==22.0.0 # 이 줄을 추가합니다. (버전은 최신 안정 버전을 명시하는 것이 좋습니다.)
alembic==1.13.1 # Alembic 추가
python-dotenv==1.0.1 # .env 파일 로드를 위해 추가
# AI/ML 관련 라이브러리 (필요에 따라 추가)
# openai # OpenAI API 사용 시
# transformers # Hugging Face 모델 사용 시 (ex: BERT, GPT)
# torch # transformers 사용 시 필요 (또는 tensorflow)
# youtube-dlp # YouTube 비디오 다운로드/정보 추출 시
# beautifulsoup4 # 웹 크롤링 시
# lxml # beautifulsoup4의 파서로 사용
# unstructured # 다양한 문서 형식에서 텍스트 추출 시